Fixed the issues detailed in #757
